programs & libraries fixed

+++ homework1/include/getarg.h
@@ -8,6 +8,6 @@
 #ifndef getarg_h
 #define getarg_h
 
-int get_argument(int argc, char* argv[], double* return_value);
+int get_argument(int argc, char* argv[], int* return_value);
 
 #endif
