+++ homework1/src/lib/arg/getarg.c
@@ -6,7 +6,7 @@
    */
 #include "getarg.h"
 
-int get_argument(int argc, char* argv[], double* return_value)
+int get_argument(int argc, char* argv[], int* return_value)
 {
         if (argc <= 1)
         {
@@ -14,7 +14,7 @@ int get_argument(int argc, char* argv[], double* return_value)
         }
         else
         {
-                *return_value = atof(argv[1]);
+                *return_value = atoi(argv[1]);
                 return 0;
         }
 }
